\name{Undocumented functions}
\alias{DoConstrain}
\alias{MidBreaks}
\alias{MidBreaks.t} 
\alias{Rinv}
\alias{approx2d} 
\alias{brk2mid} 
\alias{df2var}
\alias{dgfree} 
\alias{dmt}
\alias{dmt.paired} 
\alias{dmtmix}
\alias{dmtmix.paired} 
\alias{gausei} 
\alias{lfdr}
\alias{lfdr.paired}
\alias{rmatrix}
\alias{smooth2d.basic}
\alias{smooth2d.direct}
\title{Undocumented functions}
\description{
A number of functions is currently undocumented. Please refer to the source and source comments if you feel you need to use them.
}
\keyword{internal}% at least one, from doc/KEYWORDS

